Only download the geoip database once per day
We've also switched to curl instead of wget so we will only download the
databases if they've changed compared to the versions we have.

Even with the If-Modified-Since header, we can still be rate limited, so
we do also need the logic to only download once per day.

(wget has a --timestamping option, but it does not work if you also set
the name of the file you're downloading, so we get a filename with all
the query params present).

I'm comparing the timestamp and writing the file to different places,
and then moving the downloaded file. This is in case curl tries to write
the new file before reading the old timestamp. I wasn't able to verify
if was the case or not.